nak: Use u32 for register indices
This increases the size of RegRef to 32 bits but that's fine given that
it's usually in a union with SSARef which is 128 bits.  The real
advantage is that it allows us to start treating memory as a register
type which will come in really useful for spilling.

Part-of: <https://gitlab.freedesktop.org/mesa/mesa/-/merge_requests/24998>